Implement Util::PadStringRight()

+++ src/atgui.h
@@ -17,6 +17,6 @@ namespace UI
 	bool ColorPicker(float*, bool);
 	bool ColorPicker3(float*);
 	bool ColorPicker4(float*);
-	bool ReverseCheckbox(std::string, bool*, int = 18);
+	bool ReverseCheckbox(std::string, bool*, size_t = 18);
 	bool KeyBindButton(ButtonCode_t*);
 }
